tests create, read, findAll

+++ src/main/java/fr/eni/tp/auctionapp/dal/ItemDAO.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface ItemDAO {
     void insert(Item item);
 
-    Item read(long id);
+    Item read(int id);
 
     void update(Item item);
